# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Volunteer Allocation System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Volunteer {
  id: string;
  name: string;
  email: string;
  phone?: string;
  skills: string[];
  available: boolean;
  location?: string;
  totalAssignments: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateVolunteerBody {
  name: string;
  email: string;
  phone?: string;
  skills: string[];
  available?: boolean;
  location?: string;
}

export interface UpdateVolunteerBody {
  name?: string;
  email?: string;
  phone?: string;
  skills?: string[];
  available?: boolean;
  location?: string;
}

export type EmergencyPriority =
  (typeof EmergencyPriority)[keyof typeof EmergencyPriority];

export const EmergencyPriority = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
  Critical: "Critical",
} as const;

export type EmergencyStatus =
  (typeof EmergencyStatus)[keyof typeof EmergencyStatus];

export const EmergencyStatus = {
  Active: "Active",
  Resolved: "Resolved",
} as const;

export interface Emergency {
  id: string;
  description: string;
  priority: EmergencyPriority;
  status: EmergencyStatus;
  requiredSkills: string[];
  assignedVolunteers: string[];
  reason?: string;
  estimatedResponseTime?: string;
  location?: string;
  createdAt: string;
  resolvedAt?: string | null;
}

export type CreateEmergencyBodyPriority =
  (typeof CreateEmergencyBodyPriority)[keyof typeof CreateEmergencyBodyPriority];

export const CreateEmergencyBodyPriority = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
  Critical: "Critical",
} as const;

export interface CreateEmergencyBody {
  description: string;
  priority: CreateEmergencyBodyPriority;
  location?: string;
  requiredSkills?: string[];
  assignedVolunteers?: string[];
  reason?: string;
  estimatedResponseTime?: string;
}

export interface AnalyzeEmergencyBody {
  description: string;
  location?: string;
}

export type EmergencyAnalysisPriority =
  (typeof EmergencyAnalysisPriority)[keyof typeof EmergencyAnalysisPriority];

export const EmergencyAnalysisPriority = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
  Critical: "Critical",
} as const;

export interface EmergencyAnalysis {
  priority: EmergencyAnalysisPriority;
  required_skills: string[];
  assigned_volunteers: string[];
  reason: string;
  estimated_response_time: string;
  emergencyId?: string;
}

export interface AnalyticsSummary {
  totalVolunteers: number;
  availableVolunteers: number;
  totalEmergencies: number;
  activeEmergencies: number;
  resolvedEmergencies: number;
  avgResponseTime: string;
  criticalCases: number;
}

export interface PriorityBreakdown {
  Low: number;
  Medium: number;
  High: number;
  Critical: number;
}

export interface SkillDemandItem {
  skill: string;
  count: number;
}

export type GetVolunteersParams = {
  /**
   * Search by name or skill
   */
  search?: string;
  /**
   * Filter by availability
   */
  available?: boolean;
  /**
   * Filter by skill
   */
  skill?: string;
};

export type DeleteVolunteer200 = {
  success: boolean;
};

export type ToggleVolunteerAvailabilityBody = {
  available: boolean;
};

export type GetEmergenciesParams = {
  /**
   * Filter by status
   */
  status?: GetEmergenciesStatus;
};

export type GetEmergenciesStatus =
  (typeof GetEmergenciesStatus)[keyof typeof GetEmergenciesStatus];

export const GetEmergenciesStatus = {
  Active: "Active",
  Resolved: "Resolved",
} as const;

export type UpdateEmergencyStatusBodyStatus =
  (typeof UpdateEmergencyStatusBodyStatus)[keyof typeof UpdateEmergencyStatusBodyStatus];

export const UpdateEmergencyStatusBodyStatus = {
  Active: "Active",
  Resolved: "Resolved",
} as const;

export type UpdateEmergencyStatusBody = {
  status: UpdateEmergencyStatusBodyStatus;
};

export type GetRecentEmergenciesParams = {
  limit?: number;
};